# src/tsoa/swagger.json/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"FieldError": {
				"properties": {
					"field": {
						"type": "string",
						"example": "fieldName"
					},
					"message": {
						"type": "string",
						"example": "Validation error message"
					}
				},
				"required": [
					"field",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidationErrorResponse": {
				"properties": {
					"message": {
						"type": "string",
						"example": "Validation failed"
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/FieldError"
						},
						"type": "array",
						"description": "Detailed list of field errors"
					}
				},
				"required": [
					"message",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UnauthorizedErrorResponse": {
				"properties": {
					"error": {
						"type": "string",
						"example": "Unauthorized error message"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UserResponseData": {
				"properties": {
					"uid": {
						"type": "string",
						"example": "eb23e109-3f84-4a6f-97b2-cedf2dd16f5e"
					},
					"firstName": {
						"type": "string",
						"example": "John"
					},
					"lastName": {
						"type": "string",
						"example": "Doe"
					},
					"email": {
						"type": "string",
						"example": "john.doe@example.com"
					},
					"gender": {
						"type": "string",
						"example": "male"
					},
					"dob": {
						"type": "string",
						"example": "1995-06-15"
					},
					"address": {
						"type": "string",
						"nullable": true,
						"example": "123 Street, City"
					},
					"phone": {
						"type": "string",
						"nullable": true,
						"example": "9876543210"
					},
					"title": {
						"type": "string",
						"nullable": true,
						"example": "Software Engineer"
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true,
						"example": "https://example.com/avatar.jpg"
					},
					"createdAt": {
						"type": "string",
						"example": "2025-04-25T12:34:56.789Z"
					},
					"roles": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"example": [
							"Admin",
							"Editor"
						]
					}
				},
				"required": [
					"uid",
					"firstName",
					"lastName",
					"email",
					"gender",
					"dob",
					"createdAt",
					"roles"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserResponse": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double",
						"example": 200
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/UserResponseData"
					}
				},
				"required": [
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse": {
				"properties": {
					"message": {
						"type": "string",
						"description": "A human-readable message explaining the error",
						"example": "Email already exists"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InternalErrorResponse": {
				"properties": {
					"message": {
						"type": "string",
						"example": "Internal server error"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SignupRequest": {
				"properties": {
					"firstName": {
						"type": "string",
						"example": "hero"
					},
					"lastName": {
						"type": "string",
						"example": "hero"
					},
					"email": {
						"type": "string",
						"example": "hero@gmail.com"
					},
					"password": {
						"type": "string",
						"example": "lafsU*sdf9dl"
					},
					"confirmPassword": {
						"type": "string",
						"example": "lafsU*sdf9dl"
					},
					"gender": {
						"type": "string",
						"enum": [
							"male",
							"female",
							"other"
						],
						"example": "male"
					},
					"dob": {
						"type": "string",
						"example": "2025-03-05"
					},
					"address": {
						"type": "string",
						"nullable": true,
						"example": "123 Street, City"
					},
					"phone": {
						"type": "string",
						"nullable": true,
						"example": "9876543210"
					},
					"title": {
						"type": "string",
						"nullable": true,
						"example": "Software Engineer"
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true,
						"example": "https://example.com/avatar.jpg"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"password",
					"confirmPassword",
					"gender",
					"dob"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserListResponse": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double",
						"example": 200
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserResponseData"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"statusCode",
					"message",
					"data",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserUpdateRequest": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"dob": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserResponseData_": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double",
						"example": 200
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/UserResponseData"
					}
				},
				"required": [
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RoleUpdateRequest": {
				"properties": {
					"roleUids": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"example": [
							"c289f43e-89de-42ef-8e1a-35a222b1ac12",
							"e911ffb6-0f18-4e97-b06f-bf301cba763e"
						]
					}
				},
				"required": [
					"roleUids"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.never_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"DeleteUserResponse": {
				"properties": {
					"message": {
						"type": "string",
						"example": "User deleted successfully"
					},
					"data": {
						"$ref": "#/components/schemas/Record_string.never_",
						"example": {}
					},
					"statusCode": {
						"type": "number",
						"format": "double",
						"example": 200
					}
				},
				"required": [
					"message",
					"data",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteUsersRequest": {
				"properties": {
					"uids": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"example": [
							"ea66ffbf-85dc-4236-8f22-30bea8aa5b17",
							"f5998680-6dd6-4c23-adea-db481e464c6d",
							"7bac1c8a-b026-463b-a988-ea29468010c4"
						]
					}
				},
				"required": [
					"uids"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/user/signup": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "SignUp successful",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					},
					"409": {
						"description": "Conflict - email already registered",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable Entity – validation errors",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation failed",
											"errors": [
												{
													"field": "email",
													"message": "Please Enter Email"
												}
											]
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignupRequest"
							}
						}
					}
				}
			}
		},
		"/api/user": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Users fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/user/{uid}": {
			"get": {
				"operationId": "GetUserByUid",
				"responses": {
					"200": {
						"description": "User fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "User deletion successful",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteUserResponse"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/user/{userUid}": {
			"patch": {
				"operationId": "UpdateRole",
				"responses": {
					"200": {
						"description": "Roles updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserResponseData_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized – wrong credentials",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RoleUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/api/user/uids": {
			"delete": {
				"operationId": "DeleteManyUser",
				"responses": {
					"200": {
						"description": "User deletion successful",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteUserResponse"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteUsersRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}